fix lint errors
