+++ packages/databyss-services/sources/lib/index.ts
@@ -8,5 +8,4 @@ export * from './isBookSection'
 export * from './isCurrentAuthor'
 export * from './normalizePublicationId'
 export * from './generateShortName'
-export * from './download'
 export * from './bibliography'
